Refactor json classes to separate packages based on input or output
