+++ src/main/java/simulation/BFTSimulation.java
@@ -4,9 +4,9 @@ import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import simulation.io.IoInterface;
 import simulation.io.NoIo;
-import simulation.json.QueueResultsJson;
-import simulation.json.RunConfigJson;
-import simulation.json.ValidatorResultsJson;
+import simulation.json.output.QueueResultsJson;
+import simulation.json.input.RunConfigJson;
+import simulation.json.output.ValidatorResultsJson;
 import simulation.simulator.RunConfigUtil;
 import simulation.simulator.RunResults;
 import simulation.simulator.Simulator;
