+++ src/main/java/simulation/json/input/SwitchConfigJson.java
@@ -1,4 +1,4 @@
-package simulation.json;
+package simulation.json.input;
 
 public class SwitchConfigJson {
