+++ src/main/java/simulation/json/input/ValidatorConfigJson.java
@@ -1,4 +1,4 @@
-package simulation.json;
+package simulation.json.input;
 
 public class ValidatorConfigJson {
